Create new params loader helper for re-use
